remove ExpectsInlineQos
